Coiso

+++ src/airport/Airplane.java
@@ -134,7 +134,7 @@ public class Airplane {
         this.flightsAirplane = flightsAirplane;
     }
 
-    public double getFlightCost(float airTunnel, float windVelocity, double weight) {
+    public double getFlightCost(double airTunnel, double windVelocity, double weight) {
         double altitudeDiference = (Math.abs(getCruiseAltitude() - airTunnel))/1000;
         double xConsumption = altitudeDiference * 200;
         double standardAltitude = getFuelConsumption();
